requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-filter
django-cors-headers
drf-spectacular
django-environ
requests>=2.31.0
whitenoise==6.6.0
django-model-utils>=4.0.0
Pillow>=10.0.0
croniter>=6.0.0
